Add pausing functionality
The UI actually changes the state now. One thing for further
consideration, when resuming play after pausing I have the touch
‘killed’ in the game fragment. This was to prevent the user ship from
immediately moving over to where the pause button was (in the center of
the screen). 

This does require the user to lift their finger and touch again to start
moving the ship.